requirements.txt
numpy < 2.0
tqdm
typeguard
visualdl>=2.2.0
opencv-python <= 4.6.0
PyYAML
shapely
scipy
terminaltables
Cython
pycocotools
setuptools
Pillow

# for MOT evaluation and inference
lapx
motmetrics
sklearn==0.0

# for vehicleplate in deploy/pipeline/ppvehicle
pyclipper

# for culane data augumetation
imgaug>=0.4.0

# pip install -r requirements.txt  -i https://mirror.baidu.com/pypi/simple -i https://pypi.tuna.tsinghua.edu.cn/simple
# conda env export > environment.yml
# conda env create -f environment.yml